refactor(sidebar): move sidebar to components
