+++ packages/components/sidebar/sidebar.component.scss
