+++ packages/components/sidebar/xm-sidebar.module.ts
