+++ src/styles.scss
@@ -25,7 +25,7 @@ $material-icons-font-path: '~assets/fonts/material-design-icons-v5.woff2';
 @import 'packages/alert/alert';
 
 @import 'src/app/layouts/main/main-theme';
-@import 'src/app/modules/xm-sidebar/sidebar-theme';
+@import 'packages/components/sidebar/sidebar-theme';
 // increase css priority with xm-body class
 .xm-body {
   @include xm-sidebar-theme($xm-theme);
